+++ src/src/model/Recipe.java
@@ -23,7 +23,7 @@ public class Recipe implements Serializable {
 	private int f_id;
 	private List<String> ingredient = new ArrayList<>();
 	private Double r_i_count;
-	private double unit;	// 0：個、1：g、2：枚、3：少々、4：本
+	private int unit;	// 0：個、1：g、2：枚、3：少々、4：本
 ;
 
 
@@ -31,7 +31,7 @@ public class Recipe implements Serializable {
 	public Recipe(int rec_id, String r_name, String time, String image,
 			boolean wanpan, boolean save_time, boolean microwave_oven,
 			String recipe, int cooking_expenses, int eating_out_expenses,
-			String u_id, Date r_date, int r_count, int i_id, int f_id, String ingredient, Double r_i_count, double unit) {
+			String u_id, Date r_date, int r_count, int i_id, int f_id, String ingredient, Double r_i_count, int unit) {
 		Date sqlNow = new Date(System.currentTimeMillis()); // util.date => sql.dateへの変換
 		java.util.Date utilDate = sqlNow;
 
@@ -71,7 +71,7 @@ public class Recipe implements Serializable {
 		this.i_id = -1;
 		this.ingredient.add("");
 		this.r_i_count = -1.0;
-		this.unit = 0.0;
+		this.unit = 0;
 	}
 	public Recipe(int rec_id, String u_id, Date r_date) {
 
@@ -91,7 +91,7 @@ public class Recipe implements Serializable {
 		this.i_id = -1;
 		this.ingredient.add("");
 		this.r_i_count = -1.0;
-		this.unit = 0.0;
+		this.unit = 0;
 	}
 	public Recipe() {
 		Date sqlNow = new Date(System.currentTimeMillis()); // util.date => sql.dateへの変換
@@ -113,7 +113,7 @@ public class Recipe implements Serializable {
 		this.i_id = -1;
 		this.ingredient.add("");
 		this.r_i_count = -1.0;
-		this.unit = 0.0;
+		this.unit = 0;
 	}
 	//レシピ一覧表示時の空検索用コンストラクタの追加
 	public Recipe(int rec_id) {
@@ -134,7 +134,7 @@ public class Recipe implements Serializable {
 		this.i_id = -1;
 		this.ingredient.add("");
 		this.r_i_count = -1.0;
-		this.unit = 0.0;
+		this.unit = 0;
 	}
 	// get, setメソッドの追加
 	public int getRec_id() {
@@ -273,11 +273,11 @@ public class Recipe implements Serializable {
 		this.f_id = f_id;
 	}
 
-	public double getUnit() {
+	public int getUnit() {
 		return unit;
 	}
 
-	public void setUnit(double unit) {
+	public void setUnit(int unit) {
 		this.unit = unit;
 	}
